Refactor JSON type handling by moving to aana.storage.types and removing custom_types module
